+++ FmuImporter/SilKitBridge/ExitCodes.cs
@@ -1,0 +1,11 @@
+﻿// SPDX-License-Identifier: MIT
+// Copyright (c) Vector Informatik GmbH. All rights reserved.
+
+namespace SilKit;
+
+public class ExitCodes
+{
+  public const int Success = 0;
+
+  public const int ErrorDuringLog = 3001;
+}